+++ code/solver/Solver Trieda/HeurNuno/AulaHeur.h
@@ -33,7 +33,7 @@ public:
 
 	bool intersectaHorario(HorarioAula* const horario, bool deslocavel = false, int campId = -1, int unidId = -1) const;
 	bool intersectaHorario(DateTime const &dti, DateTime const &dtf, bool deslocavel = false, int campId = -1, int unidId = -1) const;
-	int nrCreditos() const { return horarios.size();  } 
+	int nrCreditos() const { return (int)horarios.size();  } 
 
 	virtual bool operator < (const AulaHeur &other) const
 	{
